[Bifrost] Moving Payload to be an internal bifrost type
Hiding the details of Payload. This type will take a different shape in the future.
