+++ crates/bifrost/src/lib.rs
@@ -15,6 +15,7 @@ mod bifrost_admin;
 mod error;
 pub mod loglet;
 mod loglet_wrapper;
+mod payload;
 pub mod providers;
 mod read_stream;
 mod record;
